music & lx: render different labels for works than posts
Posts get author (for now!) and length; works get instrumentation and
duration. This was a bit of an annoyance to sort through: to make it
work, the `Label` type needs to implement `Into<Value>` and `Object`
explicitly:

- `Into<Value>` so that it can be used directly in a template without
  further ado.
- `Object` so that it can use `call_method` so that calling the methods
  from the template actually works.

That’s “fine”, I suppose: it’s not a design I particularly like (as I am
finding to be the case with *most* parts of minijinja’s overall design),
but it gets the job done and now I know what to do in the future.