v6(styles): iterate on typography in content
I’m not yet sure how I want to handle “components” like the formatting
of blockquotes, notes, etc. within `.content`, because they *are* scoped
to that context. Maybe a good spot for a cascade layer? Or maybe “just”
doing the old-fashioned thing of making them be `.content blockquote`
and the like.