lx: only ask for *root* CSS files, not *all* CSS files
Previously, I was solving this by filtering the glob results down to
files that did not start with `_`, but because I have moved away from
SCSS to normal CSS, using LightningCSS to optimize/bundle, and have also
renamed the files to be easier to work with (`_index.css` for “entry
points” to directories, normal `content.css` and similar for regular
files in the project), that filtering no longer makes sense. Instead, I
want to treat files in the root of `_styles` as entry points that the
bundler can chase, and *not* do duplicate work.