lx: distinguish `Post` from `Page` and update `Archive`
I realized that `Archive` really wanted to be applied only to posts, not
to pages, because an `Archive` will never include an item that does not
have a date attached. Thinking about how best to plumb that through led
me to realize that although the date is notionally a kind of metadata,
it properly distinguishes not *kinds of metadata* but *kinds of item*.
Thus, introduce a `Post` type which *includes* a `Page`, because every
`Post` has *at least* the same set of fields and thus can delegate to
the underlying `Page`.

This implementation is still a *bit* annoying in some spots, so I may
add more helper methods that directly reference the underlying fields,
but it gets the job done fairly well and avoids needing to do further
filtering in the `Archive`.