Experimentation: trying to work with this `Component` design
It basically doesn’t do what I want, unfortunately, because `minijinja`
needs the underlying type to implement `Serialize` if this is going to
work:

    use minijinja::{Environment, Error};
    use serde::Serialize;

    pub(crate) trait Component: Serialize + Sized {
        const VIEW_NAME: &'static str;

        fn view(&self, env: &Environment) -> Result<String, Error> {
           env.get_template(&Self::template())?.render(self)
        }

        fn template() -> String {
           format!("components/{}.jinja", Self::VIEW_NAME)
        }
    }

The key bit is that `.render(self)`: it will *only* work if `self` there
has the `Serialize` constraint on it, because that’s how `mininjinja`
currently approaches things. (Future versions might not have that issue,
but future versions are not today.) I still like the idea, but as things
stand today, the result is that if you invoke an item that implements
`Serialize` but does *not* implement `Object` (and, for my purposes,
specifically `Object::call` as a way of invoking `Component::view`), you
get the “native” `Serialize` implementation instead. Which is fine for
how people generally use this, but doesn’t afford the “component” model
I was reaching for. Alas.

I’ll need to rethink this.