lx: emit `light-dark`-based theme for Arborium themes
Basically, wrap the normal output from Arborium into the style I showed
in [this blog post][post]. This uses `lightningcss` (which is already a
dependency) to parse the CSS rules and emit the rules again after
merging them together. It is plausibly faster to work directly with the
`Theme` data from Arborium, because a CSS parser is a *lot* of overhead
here and I print the results directly… but I don’t need to do this often
and it’s in the ~4ms range to do the whole thing.

[post]: https://v5.chriskrycho.com/notes/using-light-dark-instead-of-a-sass-mixin/

For posterity: I used Claude Code to generate a fair bit of the initial
code here, but I rewrote vast swathes of it because its output was, as
usual, extremely and needlessly verbose and annoyingly allocation-happy,
to say nothing of its complete lack of type safety in many spots!