lx: improve safety and comments on footnote definition emit
The implementation *worked* but was quite confusing, with some needless
indirection, previously. I’m still not *especially* happy with this, but
it is much clearer and I left myself some notes for maintenance. This is
a salutary reminder: reviewing code is harder than writing it, just as I
always tell other people. The stuff Claude writes often works, but it is
also often not very well-designed (at least presently!) and the reason
it works can be surprisingly subtle. Thus, add some detailed explanation
where it was confusing, improve the code where possible

Additionally, leave myself other breadcrumbs for future work in this
area: what was clear to me a year ago took a lot of mental effort to get
back in my head, which made working with the stuff I implemented via the
LLM that much more difficult.

Finally, make some small (but meaningful) improvements to the overall
clarity of the code in this area.