lx: no more document-level state for footnote backrefs
In the previous pass, I let the details of the original implementation
(and the way Claude Code expanded on them) constrain my imagination for
how this really ought to work. As I originally thought when I started
reworking what Claude Code had spit out, there is no reason for there to
be document-level tracking of footnotes in `second_pass::State`. Instead
I just need to have the `EmittedDefinition` itself keep track of what
footnote index it is emitted for. This is simpler and cheaper *and* lets
me use identifiers like `fnref1:2` for the backrefs. These are better in
that they associate the backref with the footnote to which they belong,
and they simplify the rest of the code to boot.